pipelines/streaming/timmer_trigger/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-eventhub
azure-storage-blob
azure-identity